Ticket2Test AI/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic>=2.0.0
litellm
instructor
python-dotenv
pyyaml
